macOS and MSVC build support (#31)
Enabling a macOS smoke test job in the pipeline, similar to the Windows'
one. To both smoke test scripts, we add the unit test model and make
sure that the generated code compiles. This exposed an issue with
compiling some of our generated code with MSVC. For HDF5, we delete the
copy constructors for the "inner" types that are used to represent the
data in a way that HDF5 supports. At the same time, we were implementing
conversion operators to convert from the inner to outer types. Deep in
the expansion of xtensor templates, we were getting errors that it was
trying to use deleted copy constructors and assignment operators.

The solution is to move away from conversion operators and to instead
pass a reference to an outer type to a `ToOuter()` method and populate
it.

There were a few other minor changes that needed to be made in order to
get the code compiling on Mac and Windows.