Document protocol schema JSON and binary format (#24)
Documenting the binary format and JSON protocol schema, which is
included in the binary and HDF5 formats.

Also making two code changes:

1. The JSON schema for vectors and arrays was not simplifying the
representation of a single type (i.e. not a union) to a single value
instead of an array with one element.
2. Simplified and made more consistent the code to read and write
floating point numbers in the binary format.