Use Mambaforge for devcontainer (#37)
The latest conda version with the version of mamba we were installing in
the base environment could no longer solve for the specification in
enviroment.yml. But we were installing mamba in a way that is [no longer
recommended](https://mamba.readthedocs.io/en/latest/installation.html).

Instead, we install mamba via mambaforge and pin the version.